requirements.txt
Flask==2.3.3
numpy
Pillow
tensorflow==2.18.0